requirements.txt
notify2
chime
python_dotenv
